more Roblox-like Color3.new behavior; class newindex pre hook for FontSize (you can now set FontSize and it will just set TextSize instead); use lua_mainthread when killing thread

+++ include/engine/datatypes/color3.hpp
@@ -6,7 +6,7 @@
 
 namespace frostbyte {
 
-int pushColor(lua_State* L, double r, double g, double b);
+int pushColor(lua_State* L, float r, float g, float b);
 int pushColor(lua_State* L, Color color);
 bool lua_iscolor(lua_State* L, int narg);
 Color* lua_checkcolor(lua_State* L, int narg);
